Add files via upload
For all the questions basic solutions are provided. Input values are supposed to be taken into consideration by the person who needs a solution to a particular problem, the second portion of code isn't provided as its already mentioned in ide by hacker rank while solving the question. Basic algorithms to solve the problem are provided here.